backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
PyPDF2==3.0.1
scikit-learn==1.4.0
nltk==3.8.1
python-dotenv==1.0.0
Werkzeug==3.0.1
numpy==1.26.3